updates from tutorial review
This commit fixes some issues that I encountered when running through the tutorial.

First, I was getting permission denied when the `docker compose up`
was attempting to build the airflow container. This was because it
tried to set permissions on the `airflow/{dags,logs,plugins}`
directories.

Next, I was getting a classic "chown permission denied" error when
starting the `database` container, because the postgresql docker
entrypoint script attempts to change the ownership of the PGDATA
directory to that of the postgres default user. Changing this to have
the target of the bind mount to `/data` instead of
`/var/lib/postgresql/data` resolved the error.

Signed-off-by: Joshua Timberman <joshua.timberman@greatexpectations.io>